refactor(address-book): extract AddressTableRow + row-composition helpers
Move ~320 lines out of AddressBookClient.tsx (713 → 393 lines):

- _components/address-table-row.tsx: AddressTableRow component + formatCreatedAt
- _lib/address-book-rows.ts: buildContractRows, buildCustomRows, filterRows,
  networkForChainId, unknownChainNetwork as pure functions

AddressBookClient.tsx wraps buildContractRows/buildCustomRows/filterRows in
its existing useMemo calls. networkForChainId is re-exported from _lib so
contractInitial (still in the page) and the row builders share one impl.

All 38 AddressBookClient.test.tsx + 29 row-composition.test.ts tests pass.
No behavior changes.
